requirements.txt
pandas
discord.py
requests